# backend/openapi.yaml
openapi: 3.0.3
info:
  title: NHS Open Access Appointment Booking API
  description: |
    RESTful API for the NHS Open Access Appointment Booking System.
    
    This API enables patients to book appointments at any NHS healthcare access point
    without traditional practice registration constraints.
    
    ## Authentication
    All endpoints (except auth endpoints) require JWT Bearer token authentication.
    
    ## Rate Limiting
    - 100 requests per minute per user
    - 1000 requests per hour per user
    
  version: 1.0.0
  contact:
    name: NHS Appointment Booking Team
    email: support@nhs-appointments.uk
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://api.nhs-appointments.uk/prod
    description: Production server
  - url: https://api.nhs-appointments.uk/staging
    description: Staging server
  - url: http://localhost:5000
    description: Local development server

tags:
  - name: Authentication
    description: User authentication and registration
  - name: Appointments
    description: Appointment management operations
  - name: Patients
    description: Patient profile and medical information
  - name: Practices
    description: Healthcare practice information

security:
  - BearerAuth: []

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user account
      description: Create a new patient account with NHS number validation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - nhs_number
                - first_name
                - last_name
                - date_of_birth
              properties:
                email:
                  type: string
                  format: email
                  example: john.smith@email.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123!
                nhs_number:
                  type: string
                  pattern: '^\d{10}$'
                  example: '9434765919'
                  description: 10-digit NHS number with valid Modulus 11 check digit
                first_name:
                  type: string
                  example: John
                last_name:
                  type: string
                  example: Smith
                date_of_birth:
                  type: string
                  format: date
                  example: '1985-03-15'
                phone:
                  type: string
                  example: '07123456789'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Registration successful
                  patient_id:
                    type: string
                    example: patient-123e4567-e89b-12d3-a456-426614174000
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login to user account
      description: Authenticate user and receive JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: john.smith@email.com
                password:
                  type: string
                  format: password
                  example: SecurePass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /appointments:
    get:
      tags:
        - Appointments
      summary: Get appointments
      description: Retrieve appointments for the authenticated user or practice
      parameters:
        - name: patient_id
          in: query
          description: Filter by patient ID (staff only)
          schema:
            type: string
        - name: practice_id
          in: query
          description: Filter by practice ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by appointment status
          schema:
            type: string
            enum: [scheduled, completed, cancelled, no-show]
        - name: from_date
          in: query
          description: Start date for date range filter
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          description: End date for date range filter
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Appointments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  appointments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Appointment'
                  count:
                    type: integer
                    example: 5
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Appointments
      summary: Create a new appointment
      description: Book a new appointment at any healthcare access point
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreate'
      responses:
        '201':
          description: Appointment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Appointment created successfully
                  appointment:
                    $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /appointments/{appointment_id}:
    get:
      tags:
        - Appointments
      summary: Get appointment details
      description: Retrieve details of a specific appointment
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Appointment details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Appointments
      summary: Update appointment
      description: Update appointment details or reschedule
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentUpdate'
      responses:
        '200':
          description: Appointment updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Appointment updated successfully
                  appointment:
                    $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Appointments
      summary: Cancel appointment
      description: Cancel an existing appointment
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Appointment cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Appointment cancelled successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /patients/{patient_id}:
    get:
      tags:
        - Patients
      summary: Get patient profile
      description: Retrieve patient profile and medical information
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
        - name: include_medical_info
          in: query
          description: Include medical information
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Patient profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Patients
      summary: Update patient profile
      description: Update patient information and medical records
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdate'
      responses:
        '200':
          description: Patient profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Patient profile updated successfully
                  patient:
                    $ref: '#/components/schemas/Patient'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'

  /practices/{practice_id}:
    get:
      tags:
        - Practices
      summary: Get practice information
      description: Retrieve healthcare practice details and availability
      parameters:
        - name: practice_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Practice information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Practice'
        '404':
          $ref: '#/components/responses/NotFound'

  /practices:
    get:
      tags:
        - Practices
      summary: Search practices
      description: Search for healthcare practices by location or name
      parameters:
        - name: postcode
          in: query
          description: Search by postcode
          schema:
            type: string
        - name: radius
          in: query
          description: Search radius in miles
          schema:
            type: integer
            default: 5
        - name: name
          in: query
          description: Search by practice name
          schema:
            type: string
      responses:
        '200':
          description: Practices found
          content:
            application/json:
              schema:
                type: object
                properties:
                  practices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Practice'
                  count:
                    type: integer

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
        user_type:
          type: string
          enum: [patient, staff, admin]
        first_name:
          type: string
        last_name:
          type: string

    Appointment:
      type: object
      properties:
        appointment_id:
          type: string
          example: appt-123e4567-e89b-12d3-a456-426614174000
        patient_id:
          type: string
        practice_id:
          type: string
        appointment_type:
          type: string
          enum: [gp_consultation, nurse_appointment, blood_test, vaccination, health_check]
        appointment_date:
          type: string
          format: date
          example: '2024-03-15'
        appointment_time:
          type: string
          format: time
          example: '14:30'
        duration_minutes:
          type: integer
          example: 15
        status:
          type: string
          enum: [scheduled, completed, cancelled, no-show]
        reason:
          type: string
          example: Annual health check
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    AppointmentCreate:
      type: object
      required:
        - practice_id
        - appointment_type
        - appointment_date
        - appointment_time
        - reason
      properties:
        practice_id:
          type: string
        appointment_type:
          type: string
          enum: [gp_consultation, nurse_appointment, blood_test, vaccination, health_check]
        appointment_date:
          type: string
          format: date
        appointment_time:
          type: string
          format: time
        reason:
          type: string
        notes:
          type: string

    AppointmentUpdate:
      type: object
      properties:
        appointment_date:
          type: string
          format: date
        appointment_time:
          type: string
          format: time
        status:
          type: string
          enum: [scheduled, completed, cancelled, no-show]
        notes:
          type: string

    Patient:
      type: object
      properties:
        patient_id:
          type: string
        nhs_number:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        email:
          type: string
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        emergency_contact:
          $ref: '#/components/schemas/EmergencyContact'
        medical_info:
          $ref: '#/components/schemas/MedicalInfo'

    PatientUpdate:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        emergency_contact:
          $ref: '#/components/schemas/EmergencyContact'
        medical_info:
          $ref: '#/components/schemas/MedicalInfo'

    Address:
      type: object
      properties:
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        postcode:
          type: string
        country:
          type: string

    EmergencyContact:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string

    MedicalInfo:
      type: object
      properties:
        allergies:
          type: array
          items:
            $ref: '#/components/schemas/MedicalAllergy'
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/MedicalCondition'
        medications:
          type: array
          items:
            $ref: '#/components/schemas/Medication'
        notes:
          type: string

    MedicalAllergy:
      type: object
      properties:
        display_text:
          type: string
          example: Penicillin
        code:
          type: string
          example: '91936005'
        system:
          type: string
          example: http://snomed.info/sct
        severity:
          type: string
          enum: [mild, moderate, severe, life-threatening]
        reaction:
          type: string

    MedicalCondition:
      type: object
      properties:
        display_text:
          type: string
          example: Type 2 Diabetes
        code:
          type: string
          example: '44054006'
        system:
          type: string
          example: http://snomed.info/sct
        clinical_status:
          type: string
          enum: [active, resolved, inactive]
        onset_date:
          type: string
          format: date

    Medication:
      type: object
      properties:
        display_text:
          type: string
          example: Metformin 500mg
        code:
          type: string
        system:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        start_date:
          type: string
          format: date

    Practice:
      type: object
      properties:
        practice_id:
          type: string
        name:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        phone:
          type: string
        email:
          type: string
        opening_hours:
          type: object
        services:
          type: array
          items:
            type: string
        accepting_patients:
          type: boolean

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object

  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad Request
            message: Validation failed
            details:
              nhs_number: Invalid NHS number format

    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid or expired token

    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: Access denied

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: Appointment not found

    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Conflict
            message: Appointment slot already booked